check-new-if: don't insert before interface "lo"
In kernel 3.10, "lo" appears after eth*, which causes new eth interfaces
to be inserted before lo in the configuration. The script needs to be
rewritten for parsing config.rc *before* parsing /proc/net/dev, so that
for each interface found there, we know whether it's already known or
if it appears after one which is already known. This would cause less
reordering. For now, the workaround consists in refusing to insert a
new interface before "lo".