config: don't create a temporary file on "config get"
It just happens to be a side effect of calling do_modify() which is
designed to change the file. But creating a temporary file during
"config get" prevents the admin user from extracting config settings.
Here we limit the changes by setting the output file to /dev/null
when doing a get.