config: do not hide errors in "diff"
"config diff" calls "save-etc -d" and "restore-etc -d" but hides their
stderr and always returns zero. The problem is that if the config cannot
be loaded, nothing is reported, making the user think everything is
synchronized while in fact it can have changes that were not saved.
Let's properly report such errors now!